Separating out edge pvs to a separate source file
